updates on messages components
