+++ src/app/components/ChatBox.tsx
@@ -1,5 +1,7 @@
 'use client' ;
 import React from 'react'
+import { useState } from 'react'
+import { sendMessage } from '../../utils/APILayer'
 import '../components/chatbox.css';
 import {
   MDBContainer,
@@ -12,208 +14,27 @@ import {
   MDBTypography,
   MDBTextArea,
   MDBCardHeader,
+  MDBInput,
 } from "mdb-react-ui-kit";
 
 const chatbox = () => {
-  
+  const [email, setEmail] = useState('');
+  const [message , setMessage] = useState('');
+
+  const handleSendMess = ()=>{
+    const userData = {
+      
+    }
+    console.log(email);
+    console.log(message);
+
+  }
+
   
   return (
     <>
-      <MDBContainer fluid className="py-5 gradient-custom">
-      <MDBRow>
-        <MDBCol md="6" lg="5" xl="4" className="mb-4 mb-md-0">
-          <h5 className="font-weight-bold mb-3 text-center text-white">
-            Member
-          </h5>
-
-          <MDBCard className="mask-custom">
-            <MDBCardBody>
-              <MDBTypography listUnStyled className="mb-0">
-                <li
-                  className="p-2 border-bottom"
-                  style={{
-                    borderBottom: "1px solid rgba(255,255,255,.3) !important",
-                  }}
-                >
-                  <a
-                    href="#!"
-                    className="d-flex justify-content-between link-light"
-                  >
-                    <div className="d-flex flex-row">
-                      <img
-                        src="https://mdbcdn.b-cdn.net/img/Photos/Avatars/avatar-8.webp"
-                        alt="avatar"
-                        className="rounded-circle d-flex align-self-center me-3 shadow-1-strong"
-                        width="60"
-                      />
-                      <div className="pt-1">
-                        <p className="fw-bold mb-0">John Doe</p>
-                        <p className="small text-white">
-                          Hello, Are you there?
-                        </p>
-                      </div>
-                    </div>
-                    <div className="pt-1">
-                      <p className="small mb-1 text-white">Just now</p>
-                      <span className="badge bg-danger float-end">1</span>
-                    </div>
-                  </a>
-                </li>
-                <li className="p-2 border-bottom">
-                  <a
-                    href="#!"
-                    className="d-flex justify-content-between link-light"
-                  >
-                    <div className="d-flex flex-row">
-                      <img
-                        src="https://mdbcdn.b-cdn.net/img/Photos/Avatars/avatar-1.webp"
-                        alt="avatar"
-                        className="rounded-circle d-flex align-self-center me-3 shadow-1-strong"
-                        width="60"
-                      />
-                      <div className="pt-1">
-                        <p className="fw-bold mb-0">Danny Smith</p>
-                        <p className="small text-white">
-                          Lorem ipsum dolor sit.
-                        </p>
-                      </div>
-                    </div>
-                    <div className="pt-1">
-                      <p className="small text-whites mb-1">5 mins ago</p>
-                    </div>
-                  </a>
-                </li>
-                <li className="p-2 border-bottom">
-                  <a
-                    href="#!"
-                    className="d-flex justify-content-between link-light"
-                  >
-                    <div className="d-flex flex-row">
-                      <img
-                        src="https://mdbcdn.b-cdn.net/img/Photos/Avatars/avatar-2.webp"
-                        alt="avatar"
-                        className="rounded-circle d-flex align-self-center me-3 shadow-1-strong"
-                        width="60"
-                      />
-                      <div className="pt-1">
-                        <p className="fw-bold mb-0">Alex Steward</p>
-                        <p className="small text-white">
-                          Lorem ipsum dolor sit.
-                        </p>
-                      </div>
-                    </div>
-                    <div className="pt-1">
-                      <p className="small text-white mb-1">Yesterday</p>
-                    </div>
-                  </a>
-                </li>
-                <li className="p-2 border-bottom">
-                  <a
-                    href="#!"
-                    className="d-flex justify-content-between link-light"
-                  >
-                    <div className="d-flex flex-row">
-                      <img
-                        src="https://mdbcdn.b-cdn.net/img/Photos/Avatars/avatar-3.webp"
-                        alt="avatar"
-                        className="rounded-circle d-flex align-self-center me-3 shadow-1-strong"
-                        width="60"
-                      />
-                      <div className="pt-1">
-                        <p className="fw-bold mb-0">Ashley Olsen</p>
-                        <p className="small text-white">
-                          Lorem ipsum dolor sit.
-                        </p>
-                      </div>
-                    </div>
-                    <div className="pt-1">
-                      <p className="small text-white mb-1">Yesterday</p>
-                    </div>
-                  </a>
-                </li>
-                <li className="p-2 border-bottom">
-                  <a
-                    href="#!"
-                    className="d-flex justify-content-between link-light"
-                  >
-                    <div className="d-flex flex-row">
-                      <img
-                        src="https://mdbcdn.b-cdn.net/img/Photos/Avatars/avatar-4.webp"
-                        alt="avatar"
-                        className="rounded-circle d-flex align-self-center me-3 shadow-1-strong"
-                        width="60"
-                      />
-                      <div className="pt-1">
-                        <p className="fw-bold mb-0">Kate Moss</p>
-                        <p className="small text-white">
-                          Lorem ipsum dolor sit.
-                        </p>
-                      </div>
-                    </div>
-                    <div className="pt-1">
-                      <p className="small text-white mb-1">Yesterday</p>
-                    </div>
-                  </a>
-                </li>
-                <li className="p-2 border-bottom">
-                  <a
-                    href="#!"
-                    className="d-flex justify-content-between link-light"
-                  >
-                    <div className="d-flex flex-row">
-                      <img
-                        src="https://mdbcdn.b-cdn.net/img/Photos/Avatars/avatar-5.webp"
-                        alt="avatar"
-                        className="rounded-circle d-flex align-self-center me-3 shadow-1-strong"
-                        width="60"
-                      />
-                      <div className="pt-1">
-                        <p className="fw-bold mb-0">Lara Croft</p>
-                        <p className="small text-white">
-                          Lorem ipsum dolor sit.
-                        </p>
-                      </div>
-                    </div>
-                    <div className="pt-1">
-                      <p className="small text-white mb-1">Yesterday</p>
-                    </div>
-                  </a>
-                </li>
-                <li className="p-2">
-                  <a
-                    href="#!"
-                    className="d-flex justify-content-between link-light"
-                  >
-                    <div className="d-flex flex-row">
-                      <img
-                        src="https://mdbcdn.b-cdn.net/img/Photos/Avatars/avatar-6.webp"
-                        alt="avatar"
-                        className="rounded-circle d-flex align-self-center me-3 shadow-1-strong"
-                        width="60"
-                      />
-                      <div className="pt-1">
-                        <p className="fw-bold mb-0">Brad Pitt</p>
-                        <p className="small text-white">
-                          Lorem ipsum dolor sit.
-                        </p>
-                      </div>
-                    </div>
-                    <div className="pt-1">
-                      <p className="small text-white mb-1">5 mins ago</p>
-                      <span className="text-white float-end">
-                        <MDBIcon fas icon="check" />
-                      </span>
-                    </div>
-                  </a>
-                </li>
-              </MDBTypography>
-            </MDBCardBody>
-          </MDBCard>
-        </MDBCol>
-
-        <MDBCol md="6" lg="7" xl="8">
-          <MDBTypography listUnStyled className="text-white">
-            <li className="d-flex justify-content-between mb-4">
+       <MDBTypography listUnStyled className="text-white">
+            {/* <li className="d-flex justify-content-between mb-4">
               <img
                 src="https://mdbcdn.b-cdn.net/img/Photos/Avatars/avatar-6.webp"
                 alt="avatar"
@@ -295,11 +116,20 @@ const chatbox = () => {
             </li>
             <MDBBtn color="light" size="lg" rounded className="float-end">
               Send
+            </MDBBtn> */}
+            <li className="d-flex flex-row align-items-center mb-4">
+              <MDBIcon fas icon="envelope me-3" size='lg'/>
+              <MDBInput label='Enter Email' id='email' type='email'  onChange={(e) => setEmail(e.target.value)}/>
+            </li>
+
+            <li className="mb-3">
+              <MDBTextArea label="Message" id="textAreaExample" rows={4}  onChange={(e) => setMessage(e.target.value)}/>
+            </li>
+            
+            <MDBBtn color="light" size="lg" rounded className="float-end" onClick={handleSendMess}>
+              Send
             </MDBBtn>
           </MDBTypography>
-        </MDBCol>
-      </MDBRow>
-    </MDBContainer>
     </>
   )
 }
