--- src/app/components/Header.tsx
@@ -1,15 +1,0 @@
-import React from 'react'
-import Link from "next/link"
-
-const Header = () => {
-  return (
-    <div>
-      Header
-      <div>
-    <a href='/login'>logout</a>
-      </div>
-    </div>
-  )
-}
-
-export default Header
