--- src/app/components/SideBar.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const sidebar = () => {
-  return (
-    <div>
-      sidebar
-    </div>
-  )
-}
-
-export default sidebar
